requirements/development.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==4.2.5
django-autoslug==1.9.9
django-cors-headers==4.3.0
django-dirtyfields==1.9.2
django-filter==23.3
django-oauth-toolkit==2.3.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-rest-framework==0.1.0
django-versatileimagefield==3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-social-oauth2==2.1.3
drf-yasg==1.21.7
ecdsa==0.18.0
idna==3.4
inflection==0.5.1
jwcrypto==1.5.0
oauthlib==3.2.2
packaging==23.2
phonenumbers==8.13.22
Pillow==10.0.1
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.8.0
python-jose==3.3.0
python-magic==0.4.27
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.6
wrapt==1.16.0
